Inline ckbtc_minter_heartbeat
