+++ rs/tests/src/ckbtc/minter/mod.rs
@@ -1,4 +1,3 @@
-pub mod test_heartbeat;
 pub mod test_kyt;
 pub mod test_retrieve_btc;
 pub mod test_update_balance;
